Updated build w/ seperate file for toast.
