+++ src/nav-states.js
@@ -1,3 +1,5 @@
+import { ResumeApplicationToast } from "./ResumeApplicationToast.js";
+
 document.addEventListener("DOMContentLoaded", function () {
   // Define all the elements that will be manipulated
   const elements = {
@@ -162,129 +164,8 @@ document.addEventListener("DOMContentLoaded", function () {
       elements.navToast.style.marginTop = "0";
     }
   }
-});
 
-// Toast behavior for directing users to their offer/in-progress application
-class ResumeApplicationToast {
-  static isVisible = false;
-  static prequalDomain = null;
-
-  constructor(prequalDomain = "https://get.point.com") {
-    ResumeApplicationToast.prequalDomain = prequalDomain;
-    this.init();
-  }
-
-  async init() {
-    try {
-      const data = await ResumeApplicationToast.fetchVisitorData();
-      if (data) {
-        this.updateToast(data);
-      }
-    } catch (error) {
-      console.error("Error initializing ResumeApplicationToast:", error);
-    }
-  }
-
-  static async fetchVisitorData() {
-    try {
-      const url = ResumeApplicationToast.prequalDomain + "/api/v1/visitors";
-      const response = await fetch(url, {
-        method: "GET",
-        mode: "cors",
-        cache: "no-cache",
-        credentials: "include",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      return response.json();
-    } catch (error) {
-      console.error("Error fetching visitor data:", error);
-      return null;
-    }
-  }
-
-  static async fetchRedirectLink() {
-    const url = ResumeApplicationToast.prequalDomain + "/api/v2/estimates/resume";
-    try {
-      const response = await fetch(url, {
-        method: "POST",
-        mode: "cors",
-        cache: "no-cache",
-        credentials: "include",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      return response.json();
-    } catch (error) {
-      console.error("Error fetching redirect link:", error);
-      return null;
-    }
-  }
-
-  updateToast(visitor) {
-    if (!visitor) return;
-
-    // Find the toast elements we want to update
-    const toastElement = document.querySelector("#toast");
-    if (!toastElement) return;
-
-    // Find the links in the toast
-    const toastLinks = toastElement.querySelectorAll("a");
-    if (!toastLinks || toastLinks.length === 0) return;
-
-    let message = null;
-    let actionText = null;
-
-    if (visitor.hasActiveDocket && !visitor.hasActiveApplicantFlow) {
-      message = "Looks like you previously started an application";
-      actionText = "Check status";
-    } else if (visitor.hasActiveDocket) {
-      message = "Looks like you previously started an application";
-      actionText = "Resume Application";
-    } else if (visitor.estimateKey) {
-      message = "You have a recently created offer";
-      actionText = "View latest offer";
-    }
-
-    if (message && actionText) {
-      // Update the toast text
-      const textElements = toastElement.querySelectorAll(".text-link-soft_underline");
-      if (textElements && textElements.length > 0) {
-        textElements.forEach((element) => {
-          element.textContent = message;
-        });
-      }
-
-      // Update link destination for all links
-      toastLinks.forEach((link) => {
-        link.addEventListener("click", async (e) => {
-          e.preventDefault();
-          try {
-            const response = await ResumeApplicationToast.fetchRedirectLink();
-            if (response && response.url) {
-              window.location.assign(response.url);
-            } else {
-              console.error("No redirect URL received");
-              // Fallback to default behavior if no URL is received
-              window.location.href = link.getAttribute("href");
-            }
-          } catch (error) {
-            console.error("Error redirecting to application:", error);
-            // Fallback to default behavior on error
-            window.location.href = link.getAttribute("href");
-          }
-        });
-      });
-
-      ResumeApplicationToast.isVisible = true;
-    }
-  }
-}
-
-// Initialize the ResumeApplicationToast when the page loads
-document.addEventListener("DOMContentLoaded", function () {
+  // Initialize the ResumeApplicationToast when the page loads
   const prequalDomain = "https://get.point.com";
   const resumeApplicationToast = new ResumeApplicationToast(prequalDomain);
 });
